Implement support for TCP tunnelling server
Implement a TCP tunnelling server according to ISO 22510:2019.

In addition to TCP, the server can also listen on a Unix domain socket,
where the same protocol as for TCP will be used. A Unix domain socket
will allow using the permissions of the socket to limit who can access
the KNX bus (only when using the systemd integration).